requirement.txt
python-binance==1.0.36
requests==2.31.0
python-dotenv==1.0.0